Fixed x64 warning.

+++ src/plugin/CommonFunc.cpp
@@ -168,7 +168,7 @@ void InsertCommas(wchar_t *Dest)
 void InsertCommas(char *Dest)
 {
 	int I;
-	for (I=strlen(Dest)-4;I>=0;I-=3)
+	for (I=(int)strlen(Dest)-4;I>=0;I-=3)
 		if (Dest[I])
 		{
 			memmove(Dest+I+2,Dest+I+1,strlen(Dest+I));
